Final project
So this is a page that's created using mostly Javascript functions to
syntax the HTML code together. Everything save the video link should
work. It creates a table based on several variables in the code for rows
and columns and redraws itself should you search for a certain phrase.
A lot of the files that show up will be the linux fedora pdf file and
that's because I'm a lazy bastard that couldn't be bothered to download
25 total files.